+++ _todo/module.spreadsheet/module.platoons.js
