+++ modules/module.help/config.json
@@ -4,9 +4,9 @@
 
     "name":"Help command",
 
-    "version":"2.0",
+    "version":"2.4",
     
-    "type":"command",
+    "type":"preMonitor",
     
     "permission":"anyone",
