+++ modules/module.qmonitor/config.json
