+++ modules/module.swgoh/config.json
@@ -4,7 +4,7 @@
 
     "name":"SWGOH-Discord",
 
-    "version":"2.0",
+    "version":"3.0",
     
     "type":"command",
